+++ src/data/variables.js
@@ -33,8 +33,10 @@ var searchButton;
 var browserLanguage;
 var browserCurrency;
 var browserMetricSystem;
-
 var addButtonIcons;
+
+var previousSelection;
+
 
 
 /// Look for these words to find that selected text is address