+++ src/functions/background.js
@@ -1,13 +1,9 @@
 /// Listener to open url in new tab
 chrome.runtime.onMessage.addListener(
     function (request, sender, sendResponse) {
-        // chrome.tabs.create({
-        //     url: request.url, active: request.focused
-        // });
-        // sendResponse({ farewell: "done" });
 
         /// Open url in new tab next to current one
-        if (request.type == 'selecton-open-new-tab')
+        if (request.type == 'selecton-open-new-tab') {
             chrome.tabs.query({
                 active: true, currentWindow: true
             }, tabs => {
@@ -17,5 +13,7 @@ chrome.runtime.onMessage.addListener(
                 });
             }
             );
+            return true;
+        }
     }
 );